diting: Decommonize vendor security patch level
This should be the minimum of all vendor components, including firmware
and device specific blobs. We can assume the sm8450-common blobs to be
the latest available for any device using this tree. Thus, setting this
from the device specific trees according to the stock rom correctly
sets the patch to the minimum of the vendor blobs.

Change-Id: I873011e3abdac42d89120b8b2ef5b77f61f6a15b